Add recipe collections + separate cook log
The cook log is actually a simpler table than I had originally
anticipated.

Meanwhile, collections are a bit more robust. So, I've pulled it out
into a join table.
